Add files via upload
The "Swimming Club Annual Audit.py" Python file consists of a program for a theoretical Swimming Club that keeps track of its members, their membership types in the club on the basis of their ages, their annual fees, and whether they are paid or not for the year. Next year, if they are not paid, those members are removed from the club's system. The ages of the remaining members are also updated, and so are their membership types, annual fees, etc. This goes on and on every year. The "Swimming Club Annual Audit Access Dates List.txt" text file keeps a permanent record of all the dates on which the Annual Audit was accessed (the Python file), and the most recent date is checked against the current date to see whether it is the present year or the previous year(s), for which the updates are made.